Moved BloggingRemindersTracker to its own file.
